requirements.txt
opencv-python
werkzeug
numpy
flask
pillow
tensorflow
 flask-cors